requirements.txt
flask
python-dotenv
groq
newspaper3k
lxml[html_clean]
